HDFS-7854. Separate class DataStreamer out of DFSOutputStream. Contributed by Li Bo.
